convAE_train file migrated
